+++ src/Arguments/Extension/AzdArguments.cs
@@ -10,4 +10,13 @@ public class AzdArguments : GlobalArguments
 {
     [JsonPropertyName(ArgumentDefinitions.Extension.Azd.CommandName)]
     public string? Command { get; set; }
+
+    [JsonPropertyName(ArgumentDefinitions.Extension.Azd.CwdName)]
+    public string? Cwd { get; set; }
+
+    [JsonPropertyName(ArgumentDefinitions.Extension.Azd.EnvironmentName)]
+    public string? Environment { get; set; }
+
+    [JsonPropertyName(ArgumentDefinitions.Extension.Azd.LearnName)]
+    public bool Learn { get; set; }
 }