+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Bugs Fixed
 
 - Fixes Service Bus host name parameter description. https://github.com/Azure/azure-mcp/pull/209/
+- Updates the usage patterns of Azure Developer CLI (azd) when invoked from MCP. https://github.com/Azure/azure-mcp/pull/203
+
+### Features Added
+
+### Other Changes
 
 ## 0.0.18 (2025-05-14)
 
@@ -17,6 +22,7 @@
 ### Bugs Fixed
 
 - Added an opt-in timeout for browser-based authentication to handle cases where the process waits indefinitely if the user closes the browser. https://github.com/Azure/azure-mcp/pull/189
+
 
 ## 0.0.16 (2025-05-13)